+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 - [BUGFIX] Fix padding and typography on information screen
 - [BUGFIX] Fix icon color dropdown menu on Deleted keys
 - [CI] Update deps
-- [REFACTOR] Preview for updater card
+- [REFACTOR] Preview(Updater card, Device info)
 
 # 1.1.5 - HotFix